Makefile: use $(file) I/O in the "FORCE"'d "TRACK" templates
Provide an alternate implementation of the recently added
"TRACK_template" which uses GNU make 4.2 features. If the 'file'
function is available we don't need to shell out at all to check if
our tracked files change.

We need to use the intermediate TRACK_template calling a
TRACK_template_eval within the generated rule so that we don't eagerly
fleshen these when "make" reads the file.

This doesn't make the runtime faster on my system, but helps to cut
down on the noise of things we shell out for
unconditionally. I.e. running "make" with "SHELL_PATH='sh -x'" (twice,
so we pick up the setting) shows than a no-op run went from emitting:

    $ SHELL_PATH='sh -x' make >/dev/null 2>&1; SHELL_PATH='sh -x' make 2>&1 |wc -l
    76

To:

    $ SHELL_PATH='sh -x' make >/dev/null 2>&1; SHELL_PATH='sh -x' make 2>&1 |wc -l
    48

That 76 to 48 number is a rough approximation of how many times we
shell out. "strace -f -c" similarly shows that we went from 25643 to
23606 syscalls. So this brings us further along in the goal of making
"make" do as little as possible when it's got nothing to re-build (see
[2]).

Note also that this implementation does not suffer from the "echo"
v.s. "printf" bug fixed in a preceding commit, as we're writing the
files out with a GNU make function it'll do the quoting for us
properly.

    $ make version.o GIT_USER_AGENT='f\o\\o"\ba\\r"' ; cat GIT-USER-AGENT
    make: 'version.o' is up to date.
    f\o\\o"\ba\\r"
    $ make version.o GIT_USER_AGENT='x\o\\o"\ba\\r"' ; cat GIT-USER-AGENT
        GIT-USER-AGENT parameters (changed)
    CC version.o
    x\o\\o"\ba\\r"
    $ make version.o GIT_USER_AGENT='x\o\\o"\ba\\r"' ; cat GIT-USER-AGENT
    make: 'version.o' is up to date.
    x\o\\o"\ba\\r"

1. https://lore.kernel.org/git/874kdn1j6i.fsf@evledraar.gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>