Makefile: disable GNU make built-in wildcard rules
Override built-in rules of GNU make that use a wildcard target. This
speeds things up significantly as we don't need to stat() so many
files just in case we'd be able to retrieve their contents from RCS or
SCCS. See [1] for an old mailing list discussion about how to disable
these.

This gives us a noticeable speedup on a no-op run:

    $ hyperfine --warmup 10 -L v ,~ -p 'git checkout HEAD{v} -- shared.mak; make -j8 all' 'make -j8 all REV=HEAD{v}' -c 'git checkout HEAD -- shared.mak'
    Benchmark #1: make -j8 all REV=HEAD
      Time (mean ± σ):     100.9 ms ±   1.7 ms    [User: 94.7 ms, System: 14.5 ms]
      Range (min … max):    99.1 ms … 105.5 ms    14 runs

    Benchmark #2: make -j8 all REV=HEAD~
      Time (mean ± σ):     112.3 ms ±   0.7 ms    [User: 106.8 ms, System: 13.8 ms]
      Range (min … max):   111.2 ms … 113.8 ms    12 runs

    Summary
      'make -j8 all REV=HEAD' ran
        1.11 ± 0.02 times faster than 'make -j8 all REV=HEAD~'

1. https://lists.gnu.org/archive/html/help-make/2002-11/msg00063.html

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>