object-list: move "struct object-list" API to its own files
Move the "struct object_list" API added in 6eb8ae00d4b ([PATCH] Header
files for object parsing, 2005-04-18) to its own set of files. This
small API only has a handful of users.

The http-push.c case could probably do with using a hash or something
like a sorted "struct string_list", since it does:

	if (!object_list_contains(objects, obj))
		object_list_insert(obj, &objects);

I.e. insertions are O(n) operations for it, and for all the users we
malloc() one item at a time. But let's just move this around for now
to establish a clear boundary between this and the rest of what
remains in object.c.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>