oid-array.h: add for_each_oid_array_oid() use in fetch-pack.c
Amend code added in 9c1e657a8fd (fetch: teach independent
negotiation (no packfile), 2021-05-04) to use a simpler
for_each_oid_array_oid() macro instead of an oid_array_for_each().

The "" here is redundant, and we'll xstrdup() those as the "name" of
the object for no reason, it should really be a NULL. But let's not
change the behavior here, and instead address that in a subsequent
commit, which will prove that we don't need a "name" here, whether
that's an xstrdup'd empty string or any other name.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>