upload-pack: use "struct object_list", not "object_array"
Change the upload-pack code to use the new "struct object_list" API
instead of "struct object_array". Since none of this code cares about
the "name", "path" or "mode" fields in the "object_array" we can save
ourselves a bit of memory by moving to this lighter wrapper.

The are non-behavior changes here, but changes that are not a simple
search/replace:

 1. We now use the for_each_*() macro to iterate over the list when
    possible, instead of counting with an "int i". This also converts
    it to our new "size_t".

 2. As we needed to create a "struct object_list_item *item" for
    those iterations, reduce the scope of the variable when possible.

 3. As noted in the preceding commit we pushed a path of "" (not NULL)
    onto the "struct object_array" in code added in 9c1e657a8fd (fetch:
    teach independent negotiation (no packfile), 2021-05-04). We're now
    getting rid of those N number of xstrdup() calls.

 4. Let's plug memory leaks associated with the old "struct
    object_array" where it's easy, e.g. in the "fetch-pack.c" caller.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>