name-rev: use "string_list", not "object_array"
Use a "struct string_list" in "builtin/name-rev.c" instead of a
"struct object_array", this fixes a memory leak that we added an
UNLEAK() for in 886e1084d78 (builtin/: add UNLEAKs, 2017-10-01).

We didn't need to xstrdup() the name on the command-line, we can just
refer back to that argv element. The only thing we need is a way to
refer to the "struct object *" associated with each "argv" element,
which the "util" member of a "struct string_list" is a perfect fit
for.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>