shallow.c: use "struct object_array_bare", not "object_array"
Change code added in ed09aef06fd (support fetching into a shallow
repository, 2006-10-30) to use "struct object_array_bare" instead of a
"struct object_array". This just needs to keep track of the objects
pushed in this loop, and doesn't need the "name", "path" and "mode"
members that "struct object_array" will provide.

While we're at it let's fix an existing memory leak that was here, we
never cleared the "stack" variable.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>