parse-options API: remove OPT_ALIAS()
Remove OPT_ALIAS(), while satisfying the use-cases that brought it
into existence in a much simpler way. The reason we've got OPT_ALIAS
is detailed in 5c387428f10 (parse-options: don't emit "ambiguous
option" for aliases, 2019-04-29), i.e. wanting to have "git clone"
understand the incomplete "--recurs" form without complaining that
"--recursive" and "--recurse-submodules" are ambiguous.

Only the "clone" caller cared about that use-case, the other four,
including two recently added by me in 98e2d9d6f7d (upload-pack:
document and rename --advertise-refs, 2021-08-05), only cared about
the small amount of typing saved by re-declaring the relevant OPT_*()
argument again with mostly the same parameters.

For the more verbose caller in "clone" let's provide a macro inline to
clearly show what is and isn't the same, and for the rest duplicate
the definition inline. The description of the aliased options was
"alias of --%s", which is being carried forward here.

To satisfy the various edge cases we need to add a new
PARSE_OPT_NO_AMBIG flag (see the new documentation in parse-options.h)
to go along with the PARSE_OPT_NO_ABBREV flag added in a preceding
commit. The two are incompatible, and as explained in the updated docs
have slightly different use-cases.

We can then implement not wanting to show the aliased flag for "clone"
in show_gitcomp() by using the PARSE_OPT_NOCOMPLETE flag.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>