parse-options API: simplify OPT_ALIAS()
Remove OPT_ALIAS() while satisfying the use-case that brought it into
existence. This is mostly a code removal, but also a code
simplification.

The reason we've got OPT_ALIAS is detailed in
5c387428f10 (parse-options: don't emit "ambiguous option" for aliases,
2019-04-29), i.e. wanting to have "git clone" understand the
incomplete "--recurs" form without complaining that "--recursive" and
"--recurse-submodules" are ambiguous.

Only the "clone" caller cared about that use-case, the other four,
including two recently added by me in 98e2d9d6f7d (upload-pack:
document and rename --advertise-refs, 2021-08-05), only cared about
the small amount of typing saved by re-declaring the relevant OPT_*()
argument again with mostly the same parameters.

For the more verbose caller in "clone" let's provide a macro inline to
clearly show what is and isn't the same, and for the rest duplicate
the definition inline. The description of the aliased options was
"alias of --%s", which is being carried forward here.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>