usage.c: use strbuf API to format usage messages
Amend the strbuf API so that it can run strbuf_vaddf() and
strbuf_addf() without calling BUG(), which means that we can use the
API for the formatting implementation of the BUG() function and other
usage function.

A summary of previous significant changes to vreportf() is available
in 116d1fa6c69 (vreportf(): avoid relying on stdio buffering,
2019-10-30). This change tries to satisfy the various use cases and
edge cases outlined there and in lined commit, while avoiding the
needless API duplication.

First, teach the strbuf API a generic way to escape characters in
string in-place, and offload the code added in f2900898795 (vreport:
sanitize ASCII control chars, 2017-01-11) to this new API. We can move
the strbuf_tolower() function added in eacd6dc5941 (Add some useful
functions for strbuf manipulation., 2008-07-13) over to this new API.

What this doesn't do is satisfy the use-case of wanting to print out
an error up to the previous length of the 4096 byte buffer if malloc()
were to fail. See b5a9e435c6d (Revert "vreportf: avoid intermediate
buffer", 2017-01-11) for a mention of that.

I think that's OK in this case, because we'll continue falling back to
the 'fprintf(stderr, "BUG!!![...]' fallback added in 116d1fa6c69. We
won't be getting as pretty of a message in the case of memory
exhaustion, but we will print what got passed in to us.

[TODO: Well, no. See]

    $ git diff
    diff --git a/strbuf.c b/strbuf.c
    index ac2273e855c..51c74f8e313 100644
    --- a/strbuf.c
    +++ b/strbuf.c
    @@ -49,6 +49,7 @@ int skip_to_optional_arg_default(const char *str, const char *prefix,
      * initialized strbuf.
      */
     char strbuf_slopbuf[1];
    +int strbuf_fail;

     void strbuf_init(struct strbuf *sb, size_t hint)
     {
    @@ -90,6 +91,8 @@ void strbuf_attach(struct strbuf *sb, void *buf, size_t len, size_t alloc)
     void strbuf_grow(struct strbuf *sb, size_t extra)
     {
            int new_buf = !sb->alloc;
    +       if (strbuf_fail)
    +               die("oh no!");
            if (unsigned_add_overflows(extra, 1) ||
                unsigned_add_overflows(sb->len, extra + 1))
                    die("you want to use way too much memory");
    diff --git a/strbuf.h b/strbuf.h
    index 688331d6386..3b2ca05b7d8 100644
    --- a/strbuf.h
    +++ b/strbuf.h
    @@ -69,6 +69,7 @@ struct strbuf {
            char *buf;
     };

    +extern int strbuf_fail;
     extern char strbuf_slopbuf[];
     #define STRBUF_INIT  { .buf = strbuf_slopbuf }

    diff --git a/usage.c b/usage.c
    index 05c8311841e..1debb32fe5b 100644
    --- a/usage.c
    +++ b/usage.c
    @@ -26,6 +26,7 @@ static void vreportf(enum usage_kind kind,

            strbuf_reset(&arg);
            strbuf_reset(&out);
    +       strbuf_fail = 1;

            if (strbuf_vaddf_NOBUG(&arg, err, params) < 0)
                    goto buggy;

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>