pack-objects: avoid "potential" segfault on error
Avoid a "potential" segfault on pack-objects error, it's "potential"
because I can't think of a way to trigger it in practice, but it is
very dangerous API use.

In 339bce27f4f (builtin/pack-objects.c: add '--stdin-packs' option,
2021-02-22) we got the first user of add_pending_oid() which passed a
"name=NULL" parameter.

If the lookup from the commit-graph with a fallback to
"parse_object()" ever failed it would have potentially segfaulted in
get_reference(), which between 3c90f03d320 (Prepare git-rev-list for
tracking tag objects too, 2005-06-29) and 339bce27f4f could expect not
to be handed a "NULL" name.

Except that even that would probably not happen on most platforms,
e.g. with glibc the sprintf() is forgiving about a NULL, and would
translate it to "(null)".

But let's be careful here and handle this case, and direct any future
such caller to our add_pending_oid_no_name() wrapper. As with the
add_pending_oid_no_name() wrapper added in preceding commit this
wrapper will allow us to eventually change how all these "no name"
callers are handled internally.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>