object-array API: remove the now-unused "slopbuf" optimization
As lamented in 31faeb2088e (object_array_entry: fix memory handling of
the name field, 2013-05-25) users of the "struct object_array" API at
the time wanted various orthogonal things from it. To optimize one
common use-case the "slopbuf" was introduced to implicitly stop
xstrdup()-ing the "" string.

In preceding commits we've migrated various API users of "struct
object_array" to the "struct object_list", and finally changed the
remaining API users that provided "" as a name to handle that case
themselves.

Let's convert the few stragglers that are still relying on it to pass
an explicit NULL, which they'll now handle correctly.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>