object-array: use ALLOC_GROW()
Replace a custom allocation pattern added in 1f1e895fccc (Add "named
object array" concept, 2006-06-19) with the ALLOC_GROW() macro. At the
time that 1f1e895fccc was authored there wasn't an ALLOC_GROW() macro
yet.

This isn't done for optimization, but rather because I'm about to
create a lighter variant of the "object-array.[ch]" API, and it'll
help readability if the two are as similar as possible in this regard

This changes the allocation pattern towards more incremental
allocation. To get to ~32k items we now allocate at these numbers of
"alloc" (emitted via manual fprintf-instrumenting of ALLOC_GROW() &
the pre-image):

    24 -> 60 -> 114 -> 195 -> 316 -> 498 -> 771 -> 1180 -> 1794 -> 2715 ->
    4096 -> 6168 -> 9276 -> 13938 -> 20931 -> 31420

Whereas before we'd do:

    64 -> 192 -> 448 -> 960 -> 1984 -> 4032 -> 8128 -> 16320 -> 32704

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>