revision API: add a "no name" add_pending_object() wrapper
Add a add_pending_object_no_name() function for the common case of
wanting to pass a "" name to add_pending_object(). As shown in
31faeb2088e (object_array_entry: fix memory handling of the name
field, 2013-05-25) we map that "" implicitly to the
"object_array_slopbuf".

I'd like to get rid of that to simplify the object-array API (instead
passing NULL), consolidating these existing callers is the first step
towards that.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>