revision API: add a "no name" add_pending_object() wrapper
Add a add_pending_object_no_name() function for the common case of
wanting to pass a "" name to add_pending_object(). As shown in
31faeb2088e (object_array_entry: fix memory handling of the name
field, 2013-05-25) we map that "" implicitly to the
"object_array_slopbuf".

I'd like to get rid of that part of the API, since it's been added
we've been narrowly avoiding a segfault due to its presence (fixed in
a subsequent commit). With preceding API migrations we've that one
remaining user of it. This change helps us to reason about that
subsequent change.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>