object-array API: add and use for_each_object_array_entry()
Add a for_each_object_array_entry() macro which works like the
corresponding for_each_object_list_item() macro in the object-list
API. Use it instead of for-loops with "int i" and "unsigned i" as the
index.

It will now be easier to change the "nr" and "alloc" members of
"struct object_array" to a "size_t" and have its various API users
carry those types forward, without changing as much code as we
otherwise would have.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>