bundle: don't make a copy of "struct rev_info *"
Amend code added in 5bb0fd2cab5 (bundle: arguments can be read from
stdin, 2021-01-11) to stop making a copy of the entirety of "struct
rev_info *", and to instead make a copy of its "struct object_array
pending".

Doing so is a much less invasive change now that we have a
release_revisions(), we don't want to have to worry about freeing the
memory twice, or the details about what is or isn't a deep copy.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>